add upload and downlado files

+++ attachments/src/main/java/attachments/repository/AttachmentRepository.java
@@ -3,5 +3,5 @@ package attachments.repository;
 import attachments.model.Attachment;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AttachmentRepository extends CrudRepository<Attachment, Long> {
+public interface AttachmentRepository extends CrudRepository<Attachment, Integer> {
 }
